+++ rpc/backend/filters_test.go
@@ -6,7 +6,7 @@ import (
 	tmtypes "github.com/cometbft/cometbft/types"
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	evmtypes "github.com/evmos/ethermint/x/evm/types"
+	evmtypes "github.com/zeta-chain/ethermint/x/evm/types"
 
 	"github.com/zeta-chain/zetacore/rpc/backend/mocks"
 	ethrpc "github.com/zeta-chain/zetacore/rpc/types"